fix: move data to services; inline function explanation
